#437 Loss of UDP messages if a distant equipment is not available
git-svn-id: https://svn.code.sf.net/p/tcnopen/trdp/trunk@2418 3b5a3598-5f4e-4449-9e63-bd40438bfec0

+++ trdp/src/vos/windows/vos_thread.c
@@ -513,8 +513,8 @@ EXT_DECL void vos_getTime (
         if(QueryPerformanceCounter(&curTime) && QueryPerformanceFrequency(&timeFreq))
 #endif /*__GNUC__*/
         {
-            pTime->tv_sec = curTime.QuadPart / timeFreq.QuadPart;
-            pTime->tv_usec = (curTime.QuadPart % timeFreq.QuadPart) / (timeFreq.QuadPart / (USECS_PER_MSEC * MSECS_PER_SEC));
+            pTime->tv_sec = (long) (curTime.QuadPart / timeFreq.QuadPart);
+            pTime->tv_usec = (long) ((curTime.QuadPart % timeFreq.QuadPart) / (timeFreq.QuadPart / (USECS_PER_MSEC * MSECS_PER_SEC)));
         }
         else
         {
